requirements.txt
livekit-agents[groq,deepgram,cartesia,silero]>=0.12.0
livekit>=0.17.0
livekit-api>=0.7.0
groq>=0.9.0
pandas>=2.2.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
